As for __kernel_cosf() and __kernel_sinf(), use a fairly optimal minimax polynomial for __kernel_tanf().  The old one was the double-precision polynomial with coefficients truncated to float.  Truncation is not a good way to convert minimax polynomials to lower precision.  Optimize for efficiency and use the lowest-degree polynomial that gives a relative error of less than 1 ulp.  It has degree 13 instead of 27, and happens to be 2.5 times more accurate (in infinite precision) than the old polynomial (the maximum error is 0.017 ulps instead of 0.041 ulps).
Unlike for cosf and sinf, the old accuracy was close to being inadequate
-- the polynomial for double precision has a max error of 0.014 ulps
and nearly this small an error is needed.  The new accuracy is also a
bit small, but exhaustive checking shows that even the old accuracy
was enough.  The increased accuracy reduces the maximum relative error
in the final result on amd64 -O1 from 0.9588 ulps to 0.9044 ulps.


git-svn-id: svn://svn.freebsd.org/base/head@152284 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f